+++ include/circt/Dialect/HW/ConversionPatterns.h
