+++ lib/Dialect/ESI/Passes/ESILowerTypes.cpp
@@ -14,8 +14,8 @@
 
 #include "circt/Dialect/ESI/ESIOps.h"
 #include "circt/Dialect/ESI/ESITypes.h"
+#include "circt/Dialect/HW/ConversionPatterns.h"
 #include "circt/Dialect/HW/HWOpInterfaces.h"
-#include "circt/Support/ConversionPatterns.h"
 
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/Interfaces/ControlFlowInterfaces.h"
