+++ lib/Support/CMakeLists.txt
@@ -8,7 +8,6 @@ set_source_files_properties("${VERSION_CPP}" PROPERTIES GENERATED TRUE)
 add_circt_library(CIRCTSupport
   APInt.cpp
   BackedgeBuilder.cpp
-  ConversionPatterns.cpp
   CustomDirectiveImpl.cpp
   FieldRef.cpp
   JSON.cpp
